WIP on Note keeper app and restructed previous work intorespective folders
