+++ src/components/note/CreateNote.jsx
@@ -1,0 +1,10 @@
+import React from "react";
+
+const CreateNote = () => {
+	return (
+		<>
+			<p>CreateNote</p>
+		</>
+	);
+};
+export default CreateNote;
