+++ src/components/note/Footer.jsx
@@ -1,0 +1,10 @@
+import React from "react";
+
+const Footer = () => {
+	return (
+		<>
+			<p>Footer</p>
+		</>
+	);
+};
+export default Footer;
