+++ src/components/note/Header.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+const Header = () => {
+	return (
+		<>
+			<div className="header">
+				<img src="" alt="logo" width="70" height="50" />
+				<h1>Notes</h1>
+			</div>
+		</>
+	);
+};
+export default Header;
