+++ src/components/note/Note.jsx
@@ -1,0 +1,10 @@
+import React from "react";
+
+const Note = () => {
+	return (
+		<>
+			<p>Note</p>
+		</>
+	);
+};
+export default Note;
