+++ src/components/slot/Slotmach.jsx
