+++ src/components/todo/index.css
@@ -1,0 +1,141 @@
+@import url("https://fonts.googleapis.com/css2?family=Roboto&display=swap");
+@import url("https://fonts.googleapis.com/css2?family=Raleway&display=swap");
+@import url("https://fonts.googleapis.com/css2?family=Josefin+Sans&display=swap");
+
+* {
+	margin: 0;
+	padding: 0;
+	box-sizing: border-box;
+	background-color: #d2dbdd;
+	font-family: "Josefin Sans", "Arial Narrow", Arial sans-serif;
+	/* font-family: "Roboto", sans-serif; */
+}
+/* ToDo List */
+
+.main_div {
+	width: 100%;
+	height: 100vh;
+	background: #2ed573;
+	display: flex;
+	justify-content: center;
+	flex-direction: row;
+	align-items: center;
+	text-align: center;
+}
+
+.center_div {
+	width: 25%;
+	height: 80vh;
+	background: #f4f4f4;
+	box-shadow: 5px 5px 25px -5px rgba(0, 0, 0, 0.5);
+	border-radius: 15px;
+}
+span {
+	color: #8566aa;
+}
+h1 {
+	color: white;
+	background: transparent;
+	background-color: #8566aa;
+	padding: 6px 0 2px 0;
+	margin-bottom: 10px;
+	text-shadow: 0 2px 2px black;
+}
+
+input {
+	text-align: center;
+	height: 30px;
+	top: 10px;
+	border: none;
+	background: transparent;
+	font-size: 20px;
+	font-weight: 500;
+	width: 60%;
+	border-bottom: 2px solid #8566aa;
+	outline: none;
+}
+
+.todo_button {
+	min-height: 40px;
+	max-width: 40px;
+	min-width: 20px !important;
+	border-radius: 50% !important;
+	background-color: #8566aa !important;
+	color: white !important;
+	margin-left: 10px !important;
+	box-shadow: 5px 5px 15px -5px rgba(0, 0, 0, 0.3);
+}
+
+.todo_button span {
+	color: #fff !important;
+}
+.todo_button:hover {
+	background-color: #20bf6b !important;
+}
+
+ol {
+	margin-top: 30px;
+}
+
+ol li {
+	padding-left: 0px;
+	text-align: left;
+	font-size: 20px;
+	font-weight: 500;
+	min-height: 40px;
+	display: flex;
+	align-items: center;
+	color: #8566aa;
+	text-transform: capitalize;
+}
+
+.todo_style {
+	display: flex;
+	flex-direction: row;
+	justify-content: left;
+	align-items: center;
+}
+
+.todo_style .fa-times {
+	width: 25px;
+	height: 25px;
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	background-color: #8566aa;
+	border-radius: 50%;
+	padding: 5px;
+	margin: 0 15px 0 35px;
+	color: aliceblue;
+	box-shadow: 5px 5px 15px -5px rgba(0, 0, 0, 0.3);
+}
+
+.todo_style .fa-times:hover {
+	background-color: crimson;
+}
+
+.todo_style .fa-times .list_style {
+	color: crimson;
+}
+
+@media (max-width: 768px) {
+	.center_hooks_div {
+		width: 55%;
+	}
+	.center_div {
+		width: 55%;
+	}
+	.todo_style .fa-times {
+		margin: 0 15px 0 15px;
+		display: flex;
+		justify-content: center;
+		align-items: center;
+	}
+	button {
+		margin-left: 5px;
+		min-height: 40px;
+		width: 40px;
+		color: #fff;
+		font-size: 30px;
+	}
+}
