+++ src/components/webseriespage/index.css
@@ -1,0 +1,83 @@
+@import url("https://fonts.googleapis.com/css2?family=Roboto&display=swap");
+@import url("https://fonts.googleapis.com/css2?family=Raleway&display=swap");
+@import url("https://fonts.googleapis.com/css2?family=Josefin+Sans&display=swap");
+
+* {
+	margin: 0;
+	padding: 0;
+	box-sizing: border-box;
+	/* background-color: #d2dbdd; */
+	/* font-family: "Josefin Sans", "Arial Narrow", Arial sans-serif; */
+	font-family: "Roboto", sans-serif;
+}
+
+.cards {
+	width: 100%;
+	height: auto;
+}
+
+.card {
+	margin: 6%;
+	transition: all 0.4s cubic-bezier(0.175, 0.885, 0, 1);
+	background-color: white;
+	width: 21.25%;
+	border-radius: 12px;
+	box-shadow: 0px 14px 10px -7px rgba(0, 0, 0, 0.1);
+	float: left;
+}
+
+.card:hover {
+	box-shadow: 0px 30px 18px -8px rgba(0, 0, 0, 0.1);
+	transform: scale(1.05, 1.05);
+}
+
+.card__img {
+	width: 100%;
+	height: 235px;
+	border-top-left-radius: 12px;
+	border-top-right-radius: 12px;
+}
+
+.card__info {
+	background-color: #fff;
+	border-bottom-left-radius: 12px;
+	border-bottom-right-radius: 12px;
+	padding: 16px 24px 24px 24px;
+}
+
+.card__category {
+	text-transform: uppercase;
+	font-size: 13px;
+	letter-spacing: 2px;
+	font-weight: 500;
+	color: #868686;
+	background-color: #fff;
+}
+
+.card__title {
+	margin-top: 5px;
+	margin-bottom: 10px;
+	font-family: "cursive";
+	background-color: white;
+	text-transform: capitalize;
+}
+
+button {
+	padding: 5px 8px;
+	text-transform: uppercase;
+	font-size: 13px;
+	font-weight: 500;
+	outline: none;
+	border: 1px solid black;
+	background-color: #fff;
+	color: #868686;
+	cursor: pointer;
+}
+
+@media (max-width: 768px) {
+	.card {
+		float: none;
+		margin: 5%;
+		width: 90%;
+	}
+}
